agent/requirements.txt
strands-agents>=0.1.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
boto3>=1.34.0
